Sökområde: Göteborg
Now there's a circular available search area for Gothenburg only. Aka you will only get (and be able to choose) gothenburg (or close area) addresses.
The mid point cordinate is Göteborg C and at the moment the circle radius is 11 km. That is enough to for example reach Partille, but not Landvetter.

Also devided the text string up in shorter sections to make the code line shorter and easier to read.

Tested by trying to input addresses outside the circular area, and then it could not find the addresses. Also if we now for example typ in "Stockholm" only gothenburg addresses show up, like "Stockholmsgatan, Göteborg, VG" etc.